+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 
   - Save keystore to private user dir, refs #141
+  - Unit test polling service, refs #107
 
 
 ## [v2019.0426]